Add support for criteria-level query base
This change allows for the addition of a criteria-level query base modifier.

To add a modifier to a definition file, add a the special surveyor_query key
within the application definition as follows:

    "net": {
        "surveyor_query": ["-cmdline:TPAutoConnSvc"],
        "process_name": ["net.exe"]
    },

Doing so will append this query term to all queries that are built withing the
application stanza (in this case, within "net").

NOTE: There can be only one surveyor_query value, which must be a list with a
single item. The list item may be a query with multiple terms in a single string.